Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 int main() {
     int a; int b; int c;
     scanf("%d %d %d", &a, &b);
-    float p=a*b*c;
+    float p=(float)(a*b*c);
     printf("%.2f", p);
     return 0;
 }
